python_service/product_recommender/requirements.txt
numpy==1.26.4
pandas==2.2.2
redis==5.0.8
faiss-cpu==1.8.0
gensim==4.3.3
sentence-transformers==3.0.1
scikit-learn==1.5.1
sqlalchemy==2.0.31
pymysql==1.1.1
torch==2.3.1
transformers==4.39.3
fastapi==0.111.1
uvicorn==0.30.3
pydantic==2.8.2
surprise==0.1
kafka-python==2.0.2
